third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-source-media-expected.txt

This is a testharness.js-based test.
FAIL the <source> media attribute has no effect assert_equals: expected 2 but got 3
Harness: the test ran to completion.
